+++ dodola/services.py
@@ -10,6 +10,7 @@ from dodola.core import (
     xesmf_regrid,
     standardize_gcm,
     xclim_remove_leapdays,
+    apply_downscaling,
     qdm_rollingyearwindow,
     train_quantiledeltamapping,
     adjust_quantiledeltamapping_year,
@@ -167,7 +168,68 @@ def bias_correct(x, x_train, train_variable, y_train, out, out_variable, method)
 
 
 @log_service
-def build_weights(x, method, target_resolution=1.0, outpath=None):
+def downscale(
+    x,
+    y_climo_coarse,
+    y_climo_fine,
+    out,
+    train_variable,
+    out_variable,
+    method,
+    domain_file,
+    adjustmentfactors=None,
+    weights_path=None,
+):
+    """Downscale bias corrected model data with IO to storage
+
+    Parameters
+    ----------
+    x : str
+        Storage URL to bias corrected input data to downscale.
+    y_climo_coarse : str
+        Storage URL to input coarse-res obs climatology to use for computing adjustment factors.
+    y_climo_fine : str
+        Storage URL to input fine-res obs climatology to use for computing adjustment factors.
+    out : str
+        Storage URL to write downscaled output to.
+    adjustmentfactors : str or None, optional
+        Storage URL to write fine-resolution adjustment factors to.
+    train_variable : str
+        Variable name used in training and obs data.
+    out_variable : str
+        Variable name used as output variable name.
+    method : {"BCSD"}
+        Downscaling method to be used.
+    domain_file : str
+        Storage URL to input grid for regridding adjustment factors
+    adjustmentfactors : str, optional
+        Storage URL to write fine-resolution adjustment factors to.
+    weights_path : str or None, optional
+        Storage URL for input weights for regridding
+    """
+    bc_ds = storage.read(x)
+    obs_climo_coarse = storage.read(y_climo_coarse)
+    obs_climo_fine = storage.read(y_climo_fine)
+    domain_fine = storage.read(domain_file)
+
+    adjustment_factors, downscaled_ds = apply_downscaling(
+        bc_ds,
+        obs_climo_coarse=obs_climo_coarse,
+        obs_climo_fine=obs_climo_fine,
+        train_variable=train_variable,
+        out_variable=out_variable,
+        method=method,
+        domain_fine=domain_fine,
+        weights_path=weights_path,
+    )
+
+    storage.write(out, downscaled_ds)
+    if adjustmentfactors is not None:
+        storage.write(adjustmentfactors, adjustment_factors)
+
+
+@log_service
+def build_weights(x, method, domain_file, outpath=None):
     """Generate local NetCDF weights file for regridding climate data
 
     Parameters
@@ -176,15 +238,16 @@ def build_weights(x, method, target_resolution=1.0, outpath=None):
         Storage URL to input xr.Dataset that will be regridded.
     method : str
         Method of regridding. Passed to ``xesmf.Regridder``.
-    target_resolution : float, optional
-        Decimal-degree resolution of global grid to regrid to.
+    domain_file : str
+        Storage URL to input xr.Dataset domain file to regrid to.
     outpath : optional
         Local file path name to write regridding weights file to.
     """
     ds = storage.read(x)
-    build_xesmf_weights_file(
-        ds, method=method, target_resolution=target_resolution, filename=outpath
-    )
+
+    ds_domain = storage.read(domain_file)
+
+    build_xesmf_weights_file(ds, ds_domain, method=method, filename=outpath)
 
 
 @log_service