+++ HISTORY.rst
@@ -5,6 +5,8 @@ History
 
 0.X.X (XXXX-XX-XX)
 ------------------
+* Update buildweights service to add support for regridding to domain file. Not backwards compatible. (PR#67, @dgergel)
+* Add downscaling service. Currently support BCSD spatial disaggregation as implemented in scikit-downscale. (PR#65, @dgergel)
 * Add bias-correction quantile delta mapping (QDM) components to support Argo Workflows. New commands added: ``dodola find-qdm-ryw``, ``dodola train-qdm``, and ``dodola apply-qdm``. (PR#70, @brews)